Cast FD_SETSIZE to int to avoid compilation problems
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@333 a8d146d6-cc15-0410-8900-af154a0219e0

+++ client-src/sendbackup.c
@@ -1157,7 +1157,7 @@ start_index(
   dup2(index, 1);
   dup2(mesg, 2);
   dup2(input, 3);
-  for(index = 4; index < FD_SETSIZE; index++) {
+  for(index = 4; index < (int)FD_SETSIZE; index++) {
     if (index != dbfd()) {
       close(index);
     }
